+++ src/components/detail/detaildescription/Detailquantity.tsx
@@ -1,4 +1,4 @@
-import styles from '@/styles/detail/detailquantity.module.css';
+import styles from '@/styles/detail/detaildescription/detailquantity.module.css';
 
 export default function Detailquantity() {
 	return (
