+++ src/styles/common/accordionlayout.module.css
@@ -1,0 +1,29 @@
+.accordionContainer {
+	margin: 150px 50px 50px 50px;
+}
+
+.accordionWrapper {
+	background-color: #eeeeee; /* 회색 배경색 */
+	border-radius: 16px;
+	padding: 0 16px;
+	transition: background-color 800ms cubic-bezier(0.19, 1, 0.22, 1);
+	margin-bottom: 8px;
+}
+.accordionWrapper.active {
+	background-color: #ffffff;
+	border-radius: 16px;
+	padding: 0 16px;
+	transition: background-color 800ms cubic-bezier(0.19, 1, 0.22, 1);
+	margin-bottom: 8px;
+	box-shadow: 1px 1px 20px rgba(0, 0, 0, 0.1);
+}
+.accordionTitle {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	padding: 16px 0;
+	cursor: pointer;
+}
+.accordionTitle > h2 {
+	font-size: 25px;
+}
