+++ src/styles/detail/detaildescription/detail.module.css
