+++ src/styles/detail/detaildescription/detaildescription.module.css
