+++ src/styles/detail/detaildescription/detailquantity.module.css
